Clean up extra code and refine the effort
